backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
channels==4.1.0
uvicorn[standard]==0.30.6
boto3==1.35.19
pillow==10.4.0